% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_gauge_functions.R
\name{gg_gauge_dial}
\alias{gg_gauge_dial}
\title{Dial Gauge Function}
\usage{
gg_gauge_dial(value, gMin, gMax, breaks, gColors, gName)
}
\arguments{
\item{value}{Value to be shown on the gauge}

\item{gMin}{Minimum value of the gauge}

\item{gMax}{Maximum value of the gauge}

\item{breaks}{Breaks for color areas}

\item{gColors}{Colors for each break}

\item{gName}{Name for the gauge (optional)}
}
\description{
This function returns a grid gTree representing a hdial gauge.
}
\examples{
gg_gauge_dial(1, 0, 2, breaks=c(1), gColors=c("green", "red"))


}
